app/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
redis==5.0.0
djangorestframework==3.16.1
psycopg==3.2.12
drf-spectacular==0.29.0

